+++ azure-storage-blob/azure/storage/blob/_deserialization.py
@@ -37,6 +37,7 @@ from .models import (
     BlobPrefix,
     AccountInformation,
     UserDelegationKey,
+    PathProperties,
 )
 from ._encryption import _decrypt_blob
 from azure.storage.common.models import _list
@@ -559,3 +560,12 @@ def _convert_xml_to_user_delegation_key(response):
 def _parse_continuation_token(response):
     marker = response.headers.get('x-ms-continuation')
     return marker if marker is not '' else None
+
+
+def _parse_path_permission_and_acl(response):
+    props = PathProperties()
+    props.owner = response.headers.get('x-ms-owner')
+    props.group = response.headers.get('x-ms-group')
+    props.permissions = response.headers.get('x-ms-permissions')
+    props.acl = response.headers.get('x-ms-acl')
+    return props